Move NETQASM_VERSION to lang module
