+++ .github/workflows/publish.yml
@@ -17,6 +17,14 @@ jobs:
         pip install
         build
         --user
+    - name: Set outputs
+      id: vars
+      run: |
+        echo "::set-output name=package_version::$(cat ./netqasm/__init__.py | cut -d "'" -f2)"
+    - name: Override package 📦 version
+      if: startsWith(github.ref, 'refs/tags') == 0
+      run: |
+        echo "__version__ = '${{ steps.vars.outputs.package_version }}a${{ github.run_id }}'" > ./netqasm/__init___.py
     - name: Build a binary wheel and a source tarball
       run: >-
         python -m
