+++ netqasm/__init__.py
@@ -1,5 +1,1 @@
-# Version of the package
 __version__ = "0.9.0"
-
-# Version of the language
-NETQASM_VERSION = (0, 10)
