+++ netqasm/lang/version.py
@@ -1,0 +1,1 @@
+NETQASM_VERSION = (0, 10)
